pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.1.3",
    "email-validator>=2.2.0",
    "fastapi==0.110.1",
    "motor==3.3.1",
    "passlib>=1.7.4",
    "pydantic>=2.6.4",
    "pyjwt>=2.10.1",
    "pymongo==4.5.0",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.9",
    "razorpay>=1.4.0",
    "requests>=2.31.0",
    "tzdata>=2024.2",
    "uvicorn==0.25.0",
    "websockets>=12.0",
]
